feat: fetch remote i18n
