+++ api/src/channel/types.ts
@@ -1,0 +1,8 @@
+import { SettingCreateDto } from '@/setting/dto/setting.dto';
+
+export type ChannelSetting<N extends string = string> = Omit<
+  SettingCreateDto,
+  'group' | 'weight'
+> & {
+  group: HyphenToUnderscore<N>;
+};
