+++ api/src/extensions/channels/live-chat-tester/i18n/en/title.json
@@ -1,0 +1,3 @@
+{
+  "live_chat_tester": "Live Chat Tester"
+}
