+++ api/src/extensions/channels/offline/i18n/en/title.json
@@ -1,0 +1,3 @@
+{
+  "offline": "Canal Web"
+}
