+++ api/src/extensions/channels/offline/i18n/fr/label.json
@@ -1,0 +1,23 @@
+{
+  "verification_token": "Jeton de vérification",
+  "allowed_domains": "Domaines autorisés",
+  "start_button": "Activer `Démarrer`",
+  "input_disabled": "Désactiver la saisie",
+  "persistent_menu": "Afficher le menu persistent",
+  "greeting_message": "Message de bienvenue",
+  "theme_color": "Thème du widget",
+  "theme_color_options": {
+    "orange": "Orange",
+    "red": "Rouge",
+    "green": "Vert",
+    "blue": "Bleu",
+    "dark": "Sombre"
+  },
+  "window_title": "Titre de la fenêtre de chat",
+  "avatar_url": "Avatar du chatbot (URL)",
+  "show_emoji": "Activer le sélecteur d'Emojis",
+  "show_file": "Activer l'upload de fichiers",
+  "show_location": "Activer le partage de géolocalisation",
+  "allowed_upload_size": "Taille maximale de téléchargement (en octets)",
+  "allowed_upload_types": "Types MIME autorisés pour le téléchargement (séparés par des virgules)"
+}
