+++ api/src/extensions/channels/offline/i18n/fr/title.json
@@ -1,0 +1,3 @@
+{
+  "live_chat_tester": "Testeur Live Chat"
+}
