--- frontend/public/locales/en/offline.json
@@ -1,28 +1,0 @@
-{
-  "title": {
-    "offline": "Web Channel"
-  },
-  "label": {
-    "verification_token": "Verification Token",
-    "allowed_domains": "Allowed Domains",
-    "start_button": "Enable `Get Started`",
-    "input_disabled": "Disable Input",
-    "persistent_menu": "Display Persistent Menu",
-    "greeting_message": "Greeting Message",
-    "theme_color": "Widget Theme",
-    "theme_color_options": {
-      "orange": "Orange",
-      "red": "Red",
-      "green": "Green",
-      "blue": "Blue",
-      "dark": "Dark"
-    },
-    "window_title": "Chat Window Title",
-    "avatar_url": "Chatbot Avatar URL",
-    "show_emoji": "Enable Emoji Picker",
-    "show_file": "Enable Attachment Uploader",
-    "show_location": "Enable Geolocation Share",
-    "allowed_upload_size": "Max Upload Size (in bytes)",
-    "allowed_upload_types": "Allowed Upload Mime Types (comma separated)"
-  }
-}
